Implement basic compiler parsing for arithmetic expressions.
The regular compiler and disassembler work with stack-based byte-code, while the alt versions work with register-based bytecode (they must be used in these pairs).
To compile with the alt versions, uncomment the '#define ALT' bit in 'common.h' or uncommon the '#-DALT' in the Makefile (if you're using Make to build).